requirements.txt
torch
transformers
librosa
scikit-learn
pandas
tqdm
opencv-python
numpy
torch>=2.0.0
transformers>=4.28.0
openai-whisper
librosa>=0.10.0
scikit-learn>=1.3.0
numpy>=1.23.0
sounddevice
soundfile
matplotlib
ffmpeg-python
tqdm
joblib
